IOS-6324 Refactoring
